# schemas/openapi-v1alpha1.json
{
  "openapi": "3.1.0",
  "info": {"title":"DARKSTAR local API","version":"1.0.0-alpha.1"},
  "servers": [{"url":"http://127.0.0.1:{port}","variables":{"port":{"default":"0"}}}],
  "security": [{"bearerAuth":[]}],
  "paths": {
...
    "/api/v1/projects": {
      "get": {"operationId":"listProjects","responses":{"200":{"description":"Registered project projections","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Project"}}}}},"default":{"$ref":"#/components/responses/Error"}}},
      "post": {"operationId":"registerProject","parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectRegistration"}}}},"responses":{"201":{"description":"Project registered","headers":{"Location":{"schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"default":{"$ref":"#/components/responses/Error"}}}
    },
...
    "/api/v1/work-items": {
      "get": {"operationId":"listWorkItems","parameters":[{"name":"includeDeleted","in":"query","required":false,"schema":{"type":"boolean","default":false}},{"name":"projectId","in":"query","required":false,"schema":{"type":"string","pattern":"^project_[0-9A-HJKMNP-TV-Z]{26}$"}}],"responses":{"200":{"description":"Work-item projections","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkItem"}}}}},"default":{"$ref":"#/components/responses/Error"}}},
      "post": {"operationId":"createWorkItem","parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWorkItemRequest"}}}},"responses":{"201":{"description":"Authored work item created","headers":{"Location":{"schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkItem"}}}},"default":{"$ref":"#/components/responses/Error"}}}
    },
    "/api/v1/work-items/import": {
      "post": {"operationId":"importWorkItem","parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImportWorkItemRequest"}}}},"responses":{"201":{"description":"Externally sourced work item created","headers":{"Location":{"schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkItem"}}}},"default":{"$ref":"#/components/responses/Error"}}}
    },
...
  },
  "components": {
    "securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer"}},
    "parameters": {
...
      "IdempotencyKey":{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string","minLength":8,"maxLength":128}},
...
    },
    "responses":{"Error":{"description":"Stable API error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}},
    "schemas": {
...
      "ApiError":{"$ref":"./api-error-v1alpha1.schema.json"},
...
      "ProjectRegistration":{"type":"object","additionalProperties":false,"required":["name","source"],"properties":{"name":{"type":"string","minLength":1},"source":{"type":"string","minLength":1,"description":"Existing absolute project directory visible to the local daemon."}}},
      "Project":{"type":"object","additionalProperties":false,"required":["id","name","sourceHash","status","resourceVersion","lastGlobalPosition","createdAt","updatedAt"],"properties":{"id":{"type":"string","pattern":"^project_[0-9A-HJKMNP-TV-Z]{26}$"},"name":{"type":"string","minLength":1},"sourceHash":{"type":"string","pattern":"^[0-9a-f]{64}$"},"status":{"enum":["active","archived"]},"resourceVersion":{"type":"integer","minimum":1},"lastGlobalPosition":{"type":"integer","minimum":1},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},
...
      "CreateWorkItemRequest":{"type":"object","additionalProperties":false,"required":["projectId","title"],"properties":{"projectId":{"type":"string","pattern":"^project_[0-9A-HJKMNP-TV-Z]{26}$"},"title":{"type":"string","minLength":1},"details":{"type":"string"},"evidence":{"type":"array","uniqueItems":true,"items":{"type":"string","minLength":1}},"routingIntent":{"$ref":"#/components/schemas/WorkRoutingIntent"},"priority":{"type":"integer","minimum":0,"default":0,"deprecated":true}}},
      "WorkRoutingIntent":{"oneOf":[{"$ref":"#/components/schemas/AutomaticWorkRoutingIntent"},{"$ref":"#/components/schemas/OverrideWorkRoutingIntent"}]},
      "AutomaticWorkRoutingIntent":{"type":"object","additionalProperties":false,"required":["mode"],"properties":{"mode":{"const":"automatic"}}},
      "OverrideWorkRoutingIntent":{"type":"object","additionalProperties":false,"required":["mode","workflowId"],"properties":{"mode":{"const":"override"},"workflowId":{"type":"string","minLength":1},"workflowVersion":{"type":"string","minLength":1},"entryNodeId":{"type":"string","minLength":1},"terminalNodeIds":{"type":"array","uniqueItems":true,"items":{"type":"string","minLength":1}}}},
      "ImportWorkItemRequest":{"type":"object","additionalProperties":false,"required":["projectId","sourceReference"],"properties":{"projectId":{"type":"string","pattern":"^project_[0-9A-HJKMNP-TV-Z]{26}$"},"sourceReference":{"type":"string","minLength":1},"title":{"type":"string","minLength":1},"priority":{"type":"integer","minimum":0,"default":0}}},
      "WorkItem":{"type":"object","additionalProperties":false,"required":["id","projectId","title","evidence","routingIntent","sourceHash","priority","status","resourceVersion","lastGlobalPosition","createdAt","updatedAt"],"properties":{"deletion":{"enum":["deleting","deleted"]},"id":{"type":"string","pattern":"^work_[0-9A-HJKMNP-TV-Z]{26}$"},"projectId":{"type":"string","pattern":"^project_[0-9A-HJKMNP-TV-Z]{26}$"},"title":{"type":"string","minLength":1},"details":{"type":"string"},"evidence":{"type":"array","uniqueItems":true,"items":{"type":"string","minLength":1}},"routingIntent":{"$ref":"#/components/schemas/WorkRoutingIntent"},"sourceHash":{"type":"string","pattern":"^[0-9a-f]{64}$"},"priority":{"type":"integer","minimum":0},"status":{"enum":["open","active","completed","cancelled"]},"resourceVersion":{"type":"integer","minimum":1},"lastGlobalPosition":{"type":"integer","minimum":1},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},
...
    }
  }
}